usuniecie zbednych checkow na zalogowanie
